# gpc-back-end/swagger.json
{
  "swagger": "2.0",
  "info": { 
    "title": "Swagger GPC & Console IoT",
    "description": "Documentazione API Gestionale Passi Carrabili e Console IoT",
    "version": "1.0"
  },
  "host": "gestionale-passi-carrabili.comune.bari.it",
  "basePath": "/api",
...
  "schemes": [
    "https"
  ],
  "paths": {
    "/auth/login": {
      "post": {
        "tags": ["Autenticazione"],
        "description": "Permette l'autenticazione dell'utente al sistema",
        "consumes":[
          "application/json"
        ],
        "produces":[
          "application/json"
        ],      
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "username": {
                  "example": "any",
                  "type": "string",
                  "required": true
                },
                "password": {
                  "example": "any",
                  "type": "string",
                  "required": true
                }
              }
            }
          }    
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {  
              "type": "object",  
              "properties": {     
                "auth": { "example": true },
                "token": { "example": "<<jwt_token>>" },
                "groups": { "example": "json contenente i permessi per le funzioni del gestionale in base al profilo utente connesso" },
                "username": { "example": "any" },
                "userlogged": { "example": "json contenente i dati dell'utente connesso" }
              }  
            }
          },
          "404": {
            "description": "Username o Password Errata / L'utente non è attivo o è stato cancellato",       
            "schema": {
              "type": "object",
              "properties": {
                "auth": { "example": false },
                "token": { "example": null },
                "message": { "example": "Username o Password Errata" }
              }                      
            }
          }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "tags": ["Autenticazione"],
        "description": "Permette la disconnessione dell'utente dal sistema",
        "consumes":[
          "application/json"
        ],
        "produces":[
          "application/json"
        ], 
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "type": "string",
            "example": "Bearer <<jwt_token>>"         
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "username": {
                  "example": "any",
                  "type": "string",
                  "required": true
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "properties": {
                "auth": { "example": false },
                "token": { "example": null },
                "message": { "example": "Logout utente registrato." }
              }                      
            }
          },
          "403": {
            "description": "Forbidden",
            "schema": {
              "type": "object",
              "properties": {
                "auth": { "example": false },
                "message": { "example": "No token provided." }
              }                      
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "type": "object",
              "properties": {
                "auth": { "example": false },
                "message": { "example": "Failed to authenticate token." }
              }                      
            }
          }
        }
      }
    },
    "/auth/recuperoPassword": {
      "post": {
        "tags": ["Autenticazione"],
        "description": "Effettua il reset della password nel caso in cui sia stata dimenticata",
        "consumes":[
          "application/json"
        ],
        "produces":[
          "application/json"
        ], 
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "username": {
                  "example": "any",
                  "type": "string",
                  "required": true
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "properties": {
                "message": { "example": "Nuova password inviata all'indirizzo mail associato all'account" }
              }                      
            }
          },
          "404": {
            "description": "Username errato / Errore durante la modifica della password",
            "schema": {
              "type": "object",
              "properties": {
                "err": { "example": "json contenente l'errore restituito dal sistema" },
                "message": { "example": "Username errato" }
              }                      
            }
          }
        }
      }
    },
    "/auth/getToken": {
      "post": {
        "tags": ["Autenticazione"],
        "description": "Permette di ottenere il token jwt di accesso alle API del gestionale (per cooperazione con egov)",
        "consumes":[
          "application/json"
        ],
        "produces":[
          "application/json"
        ],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "username": {
                  "example": "any",
                  "type": "string",
                  "required": true
                },
                "password": {
                  "example": "any",
                  "type": "string",
                  "required": true
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {  
              "type": "object",  
              "properties": {     
                "auth": { "example": true },
                "token": { "example": "<<jwt_token>>" },
                "username": { "example": "any" },
                "message": { "example": "Connessione avvenuta con successo" }
              }  
            }
          },
          "404": {
            "description": "Username o Password Errata",
            "schema": {
              "type": "object",
              "properties": {
                "auth": { "example": false },
                "token": { "example": null },
                "message": { "example": "Username o Password Errata" }
              }                      
            }
          }
        }
      }
    },
...
    "/v1/civico/": {
      "post": {
        "tags": ["Shared"],
        "description": "Effettua la ricerca e, se presente, restituisce l'indirizzo dal civilario comunale",
        "consumes":[
          "application/json"
        ],
        "produces":[
          "application/json"
        ],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "indirizzo": {
                  "required": true,
                  "example": "any"
                },
                "numero": {
                  "example": "any"
                },
                "esponente": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "type": "array",
            "description": "array contenente gli indirizzi trovati"
          },
          "500": {
            "description": "Errore di sistema",
            "schema": {
              "type": "array",
              "description": "array vuoto"                     
            }
          }
        }
      }
    },
...
	  "/v1/getPoi": {
      "post": {
        "tags": ["Shared"],
        "description": "Restituisce le pratiche (tutte, valide o non valide) con l'indirizzo nel range di distanza definito dal punto in cui ci si trova",
        "consumes":[
          "application/json"
        ],
        "produces":[
          "application/json"
        ],
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "type": "string",
            "example": "Bearer <<jwt_token>>"  
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "distance": {
                  "example": "any"
                },
                "lat": {
                  "example": "any"
                },
                "lon": {
                  "example": "any"
                },
                "isValid": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {  
              "type": "object",  
              "properties": {     
                "result": { "example": "array contenente i passi carrabili che rispettano i filtri impostati" }
              }  
            }
          },
          "403": {
            "description": "Forbidden",
            "schema": {
              "type": "object",
              "properties": {
                "auth": { "example": false },
                "message": { "example": "No token provided." }
              }                      
            }
          },         
          "500": {
            "description": "Errore di sistema",
            "schema": {
              "type": "object",
              "properties": {
                "auth": { "example": false },
                "err": { "example": "json contenente l'errore restituito dal sistema" },
                "message": { "example": "Failed to authenticate token. / Errore di sistema" }
              }                      
            }
          }
        }
      }
    },
...
  }
}